리팩터링: OCR 이미지 전처리 로직 분리
